@@ -1,14 +1,13 @@
 {
-  "name": "carlgunderson.com",
+  "name": "CarlGunderson.com",
   "version": "1.0.0",
   "description": "",
   "main": "js/index.js",
   "scripts": {
     "clean": "rm -r dist/*",
-    "dev": "webpack-dev-server --inline --hot --progress --colors",
-    "build": "webpack --progress --colors",
-    "styles": "node-sass sass/index.scss dist/styles.css -w --source-map-embed",
-    "watch": "parallelshell 'npm run dev' 'npm run styles'",
+    "start": "webpack-dev-server webpack.config.babel.js --hot --progress --colors --watch",
+    "build": "webpack webpack.config.babel.js --progress --colors",
+    "styles": "node-sass sass/index.scss dist/styles.css",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
@@ -22,16 +21,24 @@
   },
   "homepage": "https://github.com/carlgunderson/carlgunderson.github.io#readme",
   "devDependencies": {
-    "babel-core": "^6.7.7",
-    "babel-loader": "^6.2.4",
-    "babel-preset-es2015": "^6.6.0",
-    "css-loader": "^0.23.1",
-    "lodash": "^4.11.1",
-    "node-sass": "^3.6.0",
-    "parallelshell": "^2.0.0",
-    "sass-loader": "^3.2.0",
-    "style-loader": "^0.13.1",
-    "webpack": "^1.13.0",
-    "webpack-dev-server": "^1.14.1"
+    "@babel/core": "^7.0.0-beta.40",
+    "@babel/preset-env": "^7.0.0-beta.40",
+    "@babel/preset-react": "^7.0.0-beta.40",
+    "@babel/preset-stage-2": "^7.0.0-beta.40",
+    "@babel/register": "^7.0.0-beta.40",
+    "babel-loader": "^8.0.0-beta.0",
+    "css-loader": "^0.28.9",
+    "node-sass": "^4.7.2",
+    "sass-loader": "^6.0.6",
+    "style-loader": "^0.20.1",
+    "webpack": "^3.11.0",
+    "webpack-dev-server": "^2.11.1"
+  },
+  "dependencies": {
+    "material-ui": "^1.0.0-beta.33",
+    "react": "^16.2.0",
+    "react-dom": "^16.2.0",
+    "react-router": "^4.2.0",
+    "react-router-dom": "^4.2.2"
   }
 }
